refactor: extract board row logic into reusable mixin
