+++ packages/board/package.json
@@ -38,6 +38,7 @@
     "polymer"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/a11y-base": "24.7.0-alpha4",
     "@vaadin/component-base": "24.7.0-alpha4"
